changing the PersistenceXMLLoader to be able to handle version="2.1" persistence xml descriptors
The code that reads the persistence.xml into an object and verifies it against the target schema has been rewritten to be more generic. You can now add a new schema by just putting it's xsd schema file into the classpath. The validate method than looks into the persistence.xml file and reads the version attribute to determine which xsd to use.